### Summary

We eliminated much of the needless queries, work that was only needed for `up` direction also happening in `down`, activerecord issues with our constant cache clearing, and caching of pg logical subscriptions [1].  

- [x] Still need to test with regions and subscriptions

Results with a local database:

```
rake up:   ~100 seconds before, 42 seconds after
rake down: ~70 seconds before, 27 seconds after
```

### Timings

```
git checkout poc_faster_migrations

for z in `seq 0 4`
do
 dropdb vmdb_test
 createdb vmdb_test
 time (RAILS_ENV=test bundle exec rake db:migrate > /dev/null)
 time (RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null)
done

echo
echo

git checkout master

for z in `seq 0 4`
do
 dropdb vmdb_test
 createdb vmdb_test
 time (RAILS_ENV=test bundle exec rake db:migrate > /dev/null)
 time (RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null)
done
```


```
Already on 'poc_faster_migrations'
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  11.19s user 7.56s system 58% cpu 31.799 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  9.70s user 6.51s system 60% cpu 26.591 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  11.93s user 6.67s system 43% cpu 42.693 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  9.98s user 6.56s system 59% cpu 27.875 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  11.97s user 6.67s system 44% cpu 42.245 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  9.61s user 6.50s system 58% cpu 27.322 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  11.84s user 6.77s system 43% cpu 42.805 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  9.79s user 6.51s system 58% cpu 27.667 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  11.71s user 6.65s system 44% cpu 41.443 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  9.47s user 6.75s system 64% cpu 25.263 total


Switched to branch 'master'
Your branch is up to date with 'upstream/master'.
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  72.02s user 8.01s system 63% cpu 2:06.99 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  49.30s user 6.90s system 81% cpu 1:09.35 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  61.80s user 6.97s system 77% cpu 1:28.80 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  50.36s user 7.02s system 77% cpu 1:13.96 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  71.07s user 7.40s system 68% cpu 1:53.99 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  49.69s user 7.04s system 80% cpu 1:10.72 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  69.11s user 7.40s system 71% cpu 1:47.16 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  50.79s user 7.10s system 77% cpu 1:14.43 total
( RAILS_ENV=test bundle exec rake db:migrate > /dev/null; )  71.27s user 7.76s system 69% cpu 1:53.19 total
( RAILS_ENV=test bundle exec rake db:migrate VERSION=0 > /dev/null; )  49.91s user 7.04s system 79% cpu 1:11.43 total
```

[1] We'll still need https://github.com/ManageIQ/pg-logical_replication/pull/6 but the change here to cache subscription information makes sense as it shouldn't change once you start migrating the database.