@Fryguy 2021-06-03
Total aside, we need to find a way to get this into the manageiq-schema repo directly.  I never understood why it has to be in core.