@Fryguy 2021-06-09
I merged https://github.com/ManageIQ/pg-logical_replication/pull/6, but we'll have to cut a patch release of that before it takes affect.