@jrafanie 2021-06-11
FYI, I have tested with 3 nightly appliances first with jansa replication and then migrated to kasparov, and then migrated to my branch for this PR:
regions:
* 99 - global
* 1 - remote
* 2 - remote

### Initial setup

I checked out jansa on all 3, renaming the existing vmdb directory from the rpm, and installed needed rpms for gem installation:

```
    dnf -y group install "development tools" && \
    dnf config-manager --setopt=epel.exclude=*qpid-proton* --setopt=tsflags=nodocs --save && \
    dnf -y install \
      ansible \
      cmake \
      copr-cli \
      glibc-langpack-en \
      libcurl-devel \
      libpq-devel \
      librdkafka \
      libssh2-devel \
      libxml2-devel \
      libxslt-devel \
      nodejs \
      openssl-devel \
      platform-python-devel \
      postgresql-server \
      postgresql-server-devel \
      qpid-proton-c-devel \
      ruby-devel \
      rubygem-bundler \
      wget \
      sqlite-devel && \
    dnf clean all
```

### Preparing for replication starting on jansa

```
export RAILS_ENV=production
export DISABLE_DATABASE_ENVIRONMENT_CHECK=1
vmdb
git checkout origin/jansa
bundle check || bundle update
mkdir log
cp -n certs/v2_key.dev certs/v2_key
cp -n config/database.pg.yml config/database.yml
bundle exec rake evm:db:region -- --region XXX
bundle exec rake --trace  db:migrate
bundle exec rake --trace  db:seed
```

Where XXX is region 99 for global and 1 and 2 for each region.

### Configured replication for jansa

* Region 1 and 2: 
```
vmdb
bin/rails r "MiqRegion.replication_type= :remote"
```

* Region 99 (global):
```
bin/rails c
$ require 'miq_pglogical'
host1 = '192.168.1.x'
host2 = '192.168.1.y'
port = '5432'
user = 'root'
password = 'smartvm'
sub1 = PglogicalSubscription.new(:host => host1, :port => port, :user => user, :dbname => 'vmdb_production', :password => password)
sub2 = PglogicalSubscription.new(:host => host2, :port => port, :user => user, :dbname => 'vmdb_production', :password => password)
MiqPglogical.save_global_region([sub1, sub2], [])
```

Verified that `User.all.pluck(:id)` contains ids from reginons 99, 1 and 2 on the global 99 database.

### Removed replication before upgrade

```
psql -U root vmdb_production -h global_ipaddress -c 'delete from pg_subscription;'
psql -U root vmdb_production -h region1_ipaddress -c 'select pg_drop_replication_slot('region_1_subscription');'
psql -U root vmdb_production -h region2_ipaddress -c 'select pg_drop_replication_slot('region_2_subscription');'
```

### Upgraded to kasparov

On region 1, 2, and 99:

```
vmdb
git checkout origin/kasparov
bundle check || bundle update
```

### Configured replication for kasparov

Same instructions as "Configured replication for jansa".

### Attempted migration from global

```
vmdb
bin/rake db:migrate
```

See that it's waiting on the remote region 1:

```
 Waiting for remote region 1 to run migration 20200424183342
```

### Migrated db to kasparov on region 1

```
vmdb
bin/rake db:migrate
```

Now, region 99 shows:
```
Waiting for remote region 1 to run migration 20200424183342
Waiting for remote region 2 to run migration 20200424183342
```

### Migrated db to kasparov on region 2

```
vmdb
bin/rake db:migrate
```

Now, region 99 migrates immediately and finishes up.

### Repeat the steps for the branch of this pull request

* Removed replication
* Upgrading to origin/poc_faster_migrations
* Configured replication
* Migrate 99 waits on region1
* Migrate region 1, 99 now waits on region 2
* Migrate 2, then 99 completes