@jrafanie 2021-06-15
Thanks @Fryguy, your suggestions around using sql binds works well.  I rebased and retested with the current PR branch with similar speed results.

rake up:  40-50 seconds (from 130 seconds)
rake down to 0:  28 seconds (from 90 seconds)