@Fryguy 2021-06-16
@jrafanie Can you take your instructions in https://github.com/ManageIQ/manageiq/pull/21259#issuecomment-859754488 and publish them as a guide in ManageIQ/guides?