@jrafanie 2021-06-18
> @jrafanie Can you take your instructions in [#21259 (comment)](https://github.com/ManageIQ/manageiq/pull/21259#issuecomment-859754488) and publish them as a guide in ManageIQ/guides?

Opened https://github.com/ManageIQ/guides/pull/467